test case for new ignore sort attribute
